Add an API endpoint for the posts index.
