+++ app/Http/Controllers/API/PostController.php
@@ -2,8 +2,34 @@
 
 namespace App\Http\Controllers\API;
 
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Zend\Diactoros\Response\RedirectResponse;
+
 class PostController extends BaseController
 {
+    /**
+     * Get an index of posts.
+     *
+     * @param \Psr\Http\Message\ServerRequestInterface $request
+     * @param \Psr\Http\Message\ResponseInterface      $response
+     * @param array                                    $args Modifiers for requesting posts. Think pagination.
+     *
+     * @return \Psr\Http\Message\ResponseInterface
+     */
+    public function index(ServerRequestInterface $request, ResponseInterface $response, array $args = [])
+    {
+        $posts = $this->writedown->getService('api')->post()->index([
+            'where'      => [],
+            'pagination' => [
+                'current_page' => array_key_exists('page', $args) ? $args['page'] : 1,
+                'per_page'     => env('MAX_ITEMS', 10),
+            ],
+        ]);
+
+        return $this->apiResponse->setData($posts['data'])->respond();
+    }
+
     /**
      * Attempt to save a new post.
      *
