+++ app/Http/Controllers/Admin/PostController.php
@@ -3,35 +3,18 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\BaseController;
-use Psr\Http\Message\ResponseInterface;
-use Psr\Http\Message\ServerRequestInterface;
-use Zend\Diactoros\Response\RedirectResponse;
 
 class PostController extends BaseController
 {
     /**
-     * Retrieve all posts.
-     *
-     * @param \Psr\Http\Message\ServerRequestInterface $request
-     * @param \Psr\Http\Message\ResponseInterface      $response
-     * @param array                                    $args Modifiers for requesting posts. Think pagination.
+     * Display the post index.
      *
      * @return \Psr\Http\Message\ResponseInterface
      */
-    public function index(ServerRequestInterface $request, ResponseInterface $response, array $args = [])
+    public function index()
     {
-        $posts = $this->writedown->getService('api')->post()->index([
-            'where'      => [],
-            'pagination' => [
-                'current_page' => array_key_exists('page', $args) ? $args['page'] : 1,
-                'per_page'     => env('MAX_ITEMS', 10),
-            ],
-        ]);
-
         return $this->respond('admin/post/index.twig', [
             'csrf'  => $this->writedown->getService('csrf')->get(),
-            'meta'  => $posts['meta'],
-            'posts' => $posts['data'],
         ]);
     }
