+++ app/Http/routes.php
@@ -17,6 +17,7 @@ $router->get('/admin/login', 'Admin\AuthController::loginForm');
 
 $router->group('/api', function ($route) use ($apiCsrfMiddleware) {
     // Posts.
+    $route->get('/posts', 'API\PostController::index');
     $route->post('/posts/store', 'API\PostController::store')
         ->middleware($apiCsrfMiddleware);
